Jmol.___JmolVersion="14.32.82" // also 15.2.82
bug fix: V3000 MOL reading deleting H atoms inappropriately
bug fix: V3000 MOL should disregard any nonbonded Xx atom (as from a
label)
bug fix: SMILES from model that has been converted from 2D may not show
stereochemistry
bug fix: JNI-InChI not set up for 2D stereochemistry (wedge/hash).
bug fix: ChemDraw CDXML reader should allow for atom with invalid
valence

new feature: LOAD xxxx.mol FILTER "noHydrogen" (Or just "NOH")
 - loads the model in 2D with no additional hydrogens and no
minimization
 - allows for SMILES and InChI creation directly from the 2D model

+++ src/org/jmol/adapter/smarter/AtomIterator.java
@@ -26,11 +26,10 @@ package org.jmol.adapter.smarter;
 
 import org.jmol.api.JmolAdapter;
 import org.jmol.api.JmolAdapterAtomIterator;
-import javajs.util.BS;
 import org.jmol.script.T;
 
+import javajs.util.BS;
 import javajs.util.Lst;
-import javajs.util.P3d;
 import javajs.util.P3d;
 import javajs.util.V3d;
 
@@ -90,7 +89,7 @@ class AtomIterator implements JmolAdapterAtomIterator {
 	
 	@Override
   public int getElementNumber() {
-		return (atom.elementNumber > 0 ? atom.elementNumber : JmolAdapter
+    return (atom.elementNumber > 0 ? atom.elementNumber : JmolAdapter
 				.getElementNumber(atom.getElementSymbol()));
 	}
 
@@ -109,7 +108,7 @@ class AtomIterator implements JmolAdapterAtomIterator {
 	
 	@Override
   public double getPartialCharge() {
-		return (double) atom.partialCharge;
+		return atom.partialCharge;
 	}
 
 	
@@ -148,13 +147,13 @@ class AtomIterator implements JmolAdapterAtomIterator {
 	
 	@Override
   public double getBfactor() {
-		return (double) atom.bfactor;
+		return atom.bfactor;
 	}
 
 	
 	@Override
   public double getOccupancy() {
-		return (double) (atom.foccupancy * 100);
+		return (atom.foccupancy * 100);
 	}
